feat(java-concurrent): Add JDK specific test suites
