+++ dd-java-agent/instrumentation/java/java-concurrent/java-concurrent-25.0/src/previewTest/java/testdog/trace/instrumentation/java/concurrent/structuredconcurrency25/StructuredTaskScope25Test.java
@@ -151,20 +151,6 @@ public class StructuredTaskScope25Test extends AbstractInstrumentationTest {
     assertTraces(trace(span().root().operationName("parent")));
   }
 
-  @Test
-  void testForkIntoCancelledScopeDoesNotLeakContinuation() throws Exception {
-    var span = tracer.startSpan("test", "parent");
-    try (var ignored = tracer.activateSpan(span)) {
-      try (var scope = StructuredTaskScope.open(new CancelOnForkJoiner<>())) {
-        scope.fork(() -> task("child"));
-        scope.join();
-      }
-    }
-    span.finish();
-
-    assertTraces(trace(span().root().operationName("parent")));
-  }
-
   Void task(String name) {
     tracer.startSpan("test", name).finish();
     return null;
@@ -172,18 +158,5 @@ public class StructuredTaskScope25Test extends AbstractInstrumentationTest {
 
   Void failingTask() {
     throw new IllegalStateException("failing");
-  }
-
-  /** Cancels the scope as soon as the first subtask is forked. */
-  static final class CancelOnForkJoiner<T> implements StructuredTaskScope.Joiner<T, Void> {
-    @Override
-    public boolean onFork(StructuredTaskScope.Subtask<? extends T> subtask) {
-      return true; // Cancel the scope as soon as the first subtask is forked.
-    }
-
-    @Override
-    public Void result() {
-      return null;
-    }
   }
 }
